created modules and connected main.js to index.html
